+++ protocol/messenger_messages_order_controller_test.go
@@ -4,7 +4,7 @@ import (
 	"sort"
 	"sync"
 
-	"github.com/status-im/status-go/protocol/transport"
+	"github.com/status-im/status-go/messaging/transport"
 
 	wakutypes "github.com/status-im/status-go/waku/types"
 )
